+++ recipes/default.rb
@@ -48,19 +48,20 @@ link '/etc/init.d/activemq' do
 end
 
 ::Chef::Recipe.send(:include, Opscode::OpenSSL::Password)
-node.set_unless['activemq']['simple_auth_password'] = secure_password 
+if node['activemq']['simple_auth_password'].nil?
+  node.set['activemq']['simple_auth_password'] = secure_password 
+end
 
 template "#{activemq_home}/conf/activemq.xml" do
   source   'activemq.xml.erb'
-  mode     '0755'
+  mode     '0600'
   owner    node['activemq']['run_as_user'] 
   group    node['activemq']['run_as_user'] 
   variables(
     :broker_user => node['activemq']['simple_auth_user'],
     :broker_password => node['activemq']['simple_auth_password']
   )
-  notifies :restart, 'service[activemq]'
-#  only_if  { node['activemq']['use_default_config'] }
+  notifies :restart, 'service[activemq]', :delayed
 end
 
 template "#{activemq_home}/conf/jetty.xml" do
@@ -68,16 +69,16 @@ template "#{activemq_home}/conf/jetty.xml" do
   mode     '0755'
   owner    node['activemq']['run_as_user']
   group    node['activemq']['run_as_user']
-  notifies :restart, 'service[activemq]'
+  notifies :restart, 'service[activemq]', :delayed
   only_if  { node['activemq']['admin_console']['customize'] }
 end
 
 template "#{activemq_home}/conf/jetty-realm.properties" do
   source   'jetty-realm.properties.erb'
-  mode     '0755'
+  mode     '0700'
   owner    node['activemq']['run_as_user']
   group    node['activemq']['run_as_user']
-  notifies :restart, 'service[activemq]'
+  notifies :restart, 'service[activemq]', :delayed
   only_if  { node['activemq']['admin_console']['credentials']['customize'] }
 end
 
@@ -93,9 +94,23 @@ template '/etc/sysconfig/activemq' do
   notifies :restart, 'service[activemq]', :delayed
 end
 
-service 'activemq' do
-  supports :restart => true, :status => true
-  action   [:enable, :start]
+# Supports CentOS 7 systemctl
+systemd_service 'activemq' do
+  description 'Apache ActiveMQ'
+  after %w( network.target syslog.target )
+  install do
+    wanted_by 'multi-user.target'
+  end
+  service do
+    exec_start "#{activemq_home}/bin/linux-#{arch}/activemq start"
+    exec_stop "#{activemq_home}/bin/linux-#{arch}/activemq stop"
+    pid_file "#{activemq_home}/bin/linux-#{arch}/ActiveMQ.pid"
+    user node['activemq']['run_as_user']
+    group node['activemq']['run_as_user']
+    type 'forking'
+    :enable
+  end
+   only_if { node['platform_version'].start_with?('7.') } # systemd
 end
 
 # symlink so the default wrapper.conf can find the native wrapper library
@@ -115,20 +130,38 @@ template "#{activemq_home}/bin/linux/wrapper.conf" do
   notifies :restart, 'service[activemq]', :delayed
 end
 
+template "#{activemq_home}/conf/credentials.properties" do
+  source 'credentials.properties.erb'
+  owner    node['activemq']['run_as_user']
+  group    node['activemq']['run_as_user']
+  mode    '0600'
+  notifies :restart, 'service[activemq]', :delayed
+end
+
 template "#{activemq_home}/conf/login.config" do
   source 'login.config.erb'
+  owner    node['activemq']['run_as_user']
+  group    node['activemq']['run_as_user']
   mode    '0644'
   notifies :restart, 'service[activemq]', :delayed
 end
 
 template "#{activemq_home}/conf/users.properties" do
   source 'users.properties.erb'
+  owner    node['activemq']['run_as_user']
+  group    node['activemq']['run_as_user']
   mode    '0644'
   notifies :restart, 'service[activemq]', :delayed
 end
 
 template "#{activemq_home}/conf/groups.properties" do
   source 'groups.properties.erb'
+  owner    node['activemq']['run_as_user']
+  group    node['activemq']['run_as_user']
   mode    '0644'
   notifies :restart, 'service[activemq]', :delayed
+end
+
+service 'activemq' do
+  action   [:enable, :start]
 end